refactor(whatsrust): extract lifecycle wrappers
